pyproject.toml
[project]
name = "introagent"
version = "1.0.0"
description = "AI-powered warm intro agent — find the best intermediary and craft personalized outreach messages"
requires-python = ">=3.9"
dependencies = [
    "crewai",
    "langchain-groq",
    "langchain-google-genai",
    "pydantic>=2.0",
    "rich>=13.0",
    "questionary>=2.0",
    "playwright>=1.40",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
